server/requirements.txt
flask==3.1.3
flask-cors==6.0.2
pandas==3.0.1
scikit-learn==1.8.0
numpy==2.4.3
pymongo==4.16.0
python-dotenv==1.2.2
